[omnibox][categorical] Merges suggestion subtypes from lower-ranking dupes
This ensures searchboxstats reports the categorical subtypes when those
suggestions are deduped with higher relevance verbatim and/or local
history matches so the backend continues to decorate categorical
suggestions.

Note this does not change the AutocompleteMatch type - thus there is
still one catch: When a SEARCH_WHAT_YOU_TYPED match is selected the
`assisted_query_info` field in the ChromeSearchboxStats proto is not
reported. This is to prevent trivial suggestions from getting credit
for being selected. In local testing, if the user input matches the
query, the decorated categorical suggestion stops being suggested. This
is likely because the backend relies on the `assisted_query_info` field
to discern the selected suggestion subtypes.

I believe not making a special case for reporting `assisted_query_info`
for categorical suggestions is the right thing to do. After all it is
hard to say the decoration added any value as the user typed the entire
query. If any special case needs to be made, that ought to be in the
backend triggering logic.

Fixed: 1472335
Change-Id: Ia35bf2dd56b99f5c92f08e6b69b8827e56f1ae90
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4775365
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1183374}