Add crash key to log if the destroyed RFH is outermost frame
The current crash reports show that the NavigationRequest is not being
reset before the destruction of the RFH when the BFCache eviction
happens (rather than the FrameTree destruction).

When evicting a BFCache entry, we reset the NavigationRequest if the
RFH's outermost main frame is the same RFH that was restored from the
NavigationRequest owned by the RFH's FrameTreeNode. We are curious to
know if this unexpected case only happens when the RFH is not the
outermost main frame. This CL adds an additional crash key to record this information.

Bug: 1468984
Change-Id: I513b12bd513ab0b3986c08953e4d85f0dcbc7e36
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4776362
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Mingyu Lei <leimy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1183421}