[SecureDLs] Use secure downloads in DownloadFramePolicy tests
This CL updates dwonload frame policy tests to use HTTPS-based downloads
instead of HTTP. This is in preparation for an upcoming change that will
cause warnings to be shown on insecure warnings (and causes these tests
to fail).

As part of this change, this CL updates the test to use non-public
hostnames instead of registrable hostnames foo[.]com and bar[.]com. This
allows the HTTPS server to offer a valid certificate to Chrome.

Bug: 1352598
Change-Id: I02fb4b70d00d1baaee50ce07966a4f13416d0342
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4622096
Reviewed-by: Min Qin <qinmin@chromium.org>
Commit-Queue: Joe DeBlasio <jdeblasio@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1183428}