Allow preloads to boost the priority of in-flight requests.
This fixes a long-standing TODO where "preload" requests were not able
to increase the priority of a matching in-flight request. This includes
both <link rel=preload> as well as resources discovered by the
preload scanner.

Historically it wouldn't have been that much of an issue since
resources of a given type would always have the same priority but that
changed with support for "fetchpriority" where, for example, images
later in a document can be tagged as a higher priority than the same
resource earlier in the document.

Bug: 1467226
Change-Id: Ibcbf0d7aa3e7eb6ac26318b2d504847c4354943f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4778702
Commit-Queue: Patrick Meenan <pmeenan@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1183453}