[floss] Fix emulator doesn't show Bluetooth tile
This CL sets the default adapter as enabled on FakeFlossManagerClient
is constructed, so that the default adapter IsPresent to Chrome.

Since the adapter is present to Chrome, browser_tests become truly
using the BT functionality and some bugs are caught:
- collision_detection_utils.cc: Missing null pointer check
- fake_floss_adapter_client.h/cc: Missing mock implementations
- floss_adapter_client.h: CreateBluetoothDeviceFloss returns the
                          VendorProductInfo having leftover value

Also, the hard-coded connected device (kBondedAddress1) generates
toast on UI started, which is affecting several browser tests.
This patch makes no connected device by default, and adds a flag for
the device tests that require connected device.

BUG=b:274815204
TEST=./out/Default/chrome --enable-features=Floss

Change-Id: I5cd2f1274db24f23429635b1b1c21fe04f96f843
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4707669
Commit-Queue: Hsin-chen Chuang <chharry@google.com>
Reviewed-by: Abhishek Pandit-Subedi <abhishekpandit@google.com>
Reviewed-by: Kazuki Takise <takise@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1183455}