Fix printing browser tests for StartBasicPrintCancel
Rework the out-of-process callback handling for TestPrinterQueryOop so
that platforms which don't support OOP system print dialogs will use the
same callbacks as the in-browser tests that use TestPrinterQuery.  This
ensures that there are now callbacks in all cases when getting default
settings and when asking the user for settings.  This affects the
expectations of multiple tests which involve basic printing.

Replace the separate tests for StartBasicPrintCancel with a single test
that covers the same permutations.

This allows testing of StartBasicPrintCancel to be enabled again,
without any dependency on the system dialog being out-of-process.

Bug: 1375007, 1414968
Change-Id: Idea5304a4ec18da9588878237acf193f3f9330f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4769609
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Alan Screen <awscreen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1183460}