Added script to generate TypeScript definition files for chrome extension APIs.
This generator works well with many of the idls and json files that were tested.

It reuses the same data objects that the js_externs generator did, however instead of outputting that data into jsdoc and javascript, it is instead output as typescript definitions.

Note:

Imports are not resolvable, so they will need to be added manually
after generation if any are required.

Bug: 1203307
Change-Id: I62440c9707bc233580123feda46dad130a44446a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4679213
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Jacob Francis <francisjp@google.com>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Auto-Submit: Jacob Francis <francisjp@google.com>
Cr-Commit-Position: refs/heads/main@{#1183462}