Add metrics that counts up speculation rules prerenders per page
This CL introduces the metric that records the number of eventually received speculation rules prerender triggers per primary page changed,
classified with their eagerness.

This aims to investigate the overall latest usage amount of speculation
rules prerenders that developers actually try to use in one page, as
well as their difference on each eagerness. This is expected to be an
insight to improve prerender limits and schedulers.

Bug: 1464021
Change-Id: I88aaa17dab5e1d0dbec79a24338c3c3ef0680416
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4725822
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Lingqi Chi <lingqi@chromium.org>
Commit-Queue: Taiyo Mizuhashi <taiyo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1183477}