Improve warning comments in PrintRenderFrameHelper
Rearrange the code so warning comments are closer to the code they are
warning about. As part of this CL, also add in a missing warning and
change some pipes to backticks in the comments.

Bug: 1459437
Change-Id: Id62906f00352e801aeafbdc213a1a933be7aa8e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4727894
Reviewed-by: Alan Screen <awscreen@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1183523}