Re-add the kForceSoftwareForLowResolutions flag for testing.
In b/294894958 it was pointed out that this flag is still desired for
testing purposes, such as being able to do HW vs SW manual analysis, or
other HW tests that depend on not doing SW fallback.

This flag is enabled-by-default on all platforms except for Android.
There are no plans at the moment to experiment with the flag, so the
default values represent today's behavior.

Bug: b/294894958
Change-Id: Iff4154b4409fddf0ed87f3673b144e3498f2a04b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4777781
Reviewed-by: Ilya Nikolaevskiy <ilnik@chromium.org>
Commit-Queue: Henrik Boström <hbos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1183548}