Set is_actively_scrolling bit during touch interactions
In crrev.com/c/4690922, we began plumbing a bit to optimize rendering
during an interaction, but in that CL it was limited to scrolling. It
turns out that this is too restrictive: just before a scroll begins,
there are touch events that can cause visual updates that can also be
delayed (and that we would like to optimize).

With this change, the is_actively_scrolling name becomes inaccurate.
I've logged crbug.com/1470329 to track cleaning this up.

credit: thanks to khushalsagar@ who had originally suggested this more
general approach on the bug and who also noticed that a compositor
frame not being presented quickly on touch start which precipitated
this CL.

Bug: 1082569
Change-Id: If48bfdecef5b3ff60a46668e152564be0621bb4a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4752394
Commit-Queue: Ian Vollick <vollick@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1183580}