Include DIRECT fallback for IP Protection Proxy
The NSProxyAllowList creates the ProxyConfig used to set up the IP
Protection Proxy, and this should contain a fallback to DIRECT if the
proxy is not available. Doing so is simply a matter of appending
`,direct://` to the rules spec.

Also: minor cleanup in `network_service_proxy_allow_list*`, and update
the test in `network_service_proxy_delegate_unittest.cc` to show the
same effect.

Bug: 1471478
Change-Id: I61b98e0a4d2b5f9e03e8f6f8a2974cf9fdc99d31
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4778887
Commit-Queue: Maks Orlovich <morlovich@chromium.org>
Reviewed-by: Brianna Goldstein <brgoldstein@google.com>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Auto-Submit: Dustin Mitchell <djmitche@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1183615}