[Auto-disable Accessibility] Fix for spannable edit text events
This is a speculative fix for a crash that is sometimes appearing when
an accessibility event is sent from Chrome and no services are enabled.
With this change we move the impl back to the deprecated method for
checking accessibility state, which preserves the original logic while
still using the new AccessibilityState impl. This will hopefully stop
crashes and the edge case can be more formally addressed when the (now)
deprecated method is removed.

AX-Relnotes: N/A
Bug: 1471886, 1430202, b/265493191
Change-Id: I136963d7c6c77f0394788e151ccd0a93e49d5d3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4778969
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Reviewed-by: Tomasz Wiszkowski <ender@google.com>
Cr-Commit-Position: refs/heads/main@{#1183660}