Use fuzzy matching in FocusRingBrowserTest ChromeRefresh2023
According to the attached bug, the ChromeRefresh2023 flag affects
rendering in a slight way that is not visible to the human eye. Setting
a separate baseline didn't resolve this, likely because
ChromeRefresh2023 is not tested in CQ+1. This patch uses a fuzzy
comparator in order to allow work to continue on ChromeRefresh2023 until
it is ready to enable by default, at which time we will stop doing fuzzy
matching and replace the baseline with the ChromeRefresh2023 one because
at that point ChromeRefresh2023 will be tested in CQ+1.

Bug: 1443584
Change-Id: I37be8ef308acf7d296346150e7c71c751ff85356
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4779075
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1183665}