Tab DnD: Using intent extra to pass the tab index when moving the tab from one instance to another.
A new intent extras key EXTRA_TAB_INDEX is defined to hold the tab index value that indicates where to position the newly added tab to the list of tabs of the TabModel.

Low-Coverage-Reason: The code is only executed in environments not in the CQ. Integration test over MultiInstanceManager31 aren't counted.
Bug: b/295037141
Change-Id: If70607c7ee9fbad057da6202bdcc25461dede0a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4772019
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Sirisha Kavuluru <skavuluru@google.com>
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Gurmeet Kalra <gurmeetk@google.com>
Cr-Commit-Position: refs/heads/main@{#1183699}