Re-enable PaymentRequestEmailAndPhoneTest
Re-enable the remaining 2 disabled tests in
PaymentRequestEmailAndPhoneTest, which was already migrated away from
basic-card usage. The metrics test is updated to the correct expected
metrics with basic-card removed.

The other disabled testAddEmailAndPhoneAndPay test was previously
disabled for flakiness. The test suite is also updated to no longer use
triggerUIAndWait which was a previous source of flakiness.

Use HistogramWatcher instead of the deprecated RecordHistogram for
metrics testing.

Bug: 1331527
Change-Id: I3227d147ee4add4d6231ea00d78e298321dec144
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4778275
Reviewed-by: Peter Kotwicz <pkotwicz@chromium.org>
Commit-Queue: Nick Burris <nburris@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1183722}