[iOS] Add Unit Tests for Hash Prefix Realtime Testing
This is a follow up to crrev.com/c/4757419 we are adding tests in this
CL because we needed to discuss the proper way of testing hash prefix
realtime. These unit tests check to make sure that we're logging hash
realtime histogram metrics and that we receive the correct perform
check. The histogram log should always add a count since we log one
entry regardless of if it's true or false.

Bug: 1468377
Change-Id: I19fc9ec510aab642e8c29c2021be75c1e48e558d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4764848
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Commit-Queue: Joemer Ramos <joemerramos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1183733}