wm: Fix overview and alt tab rounding at certain aspect ratios
With some aspect ratios, we can get a scenario where the backdrop for
overview and alt tab is larger than the thumbnail, but the thumbnail's
corners stick out. To deal with this case, we update the check for
backdrop visibility to also check if such corners will stick out.

Screenshot of double rounding case:
https://screenshot.googleplex.com/4HLU4tvCDovahpG

Test: manual
Test: ash_unittest OverviewUtilsTest.ShouldRoundThumbnailWindow
Change-Id: Ie296719265453c07e36e4716f10592ee1bedada9
Fixed: b/280674084
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4778994
Reviewed-by: Connie Xu <conniekxu@chromium.org>
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1183737}