fido: abstract some Windows WebAuthn API calls
Abstract calls to WebAuthNEnumeratePlatformCredentials() behind a more
sane C++ API, so we don't need to deal with memory management for the
output pointers in multiple places.

Also change behavior of the corresponding API fake to not return an
output value on NTE_NOT_FOUND, making it match up with the real
implementation.

Change-Id: I5d1158ec4185b5f867ac0684a4e007b8ed9ad56c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4757003
Commit-Queue: Martin Kreichgauer <martinkr@google.com>
Reviewed-by: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1183739}