media/gpu/v4l2/V4L2StatefulVD: Implement Reset()
This CL adds the implementation of Reset(); this implementation
essentially:
- Cancels all in-flight operations (from the event monitoring
function)
- Flushes pending work and
- Restarts both |OUTPUT_queue_| and |CAPTURE_queue|.

An interesting problem with the weak pointer invalidation is that
we still need it to stay valid for the |client_|s frame pool to
NotifyWhenFrameAvailable(); this CL adds a new WeakPtrFactory (one
for each use), and also removes the use of |weak_this_|, preferring
instead to just GetWeakPtr()s directly.

With this CL ResetAfterInitialize, ResetEndOfStream and ReseMidStream
test cases are passing on trogdor.

Bug: b:279980150
Change-Id: I38617cc51b9f605f23a8ba476da0b6f83d678211
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4744641
Reviewed-by: Steve Cho <stevecho@chromium.org>
Reviewed-by: Miguel Casas-Sanchez <mcasas@chromium.org>
Reviewed-by: Hang Nguyen <hnt@chromium.org>
Commit-Queue: Miguel Casas-Sanchez <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1183742}