Fix custom-elements/perform-microtask-checkpoint-before-construction.html failure
See crbug.com/1420810#c9 for analysis of why we fail the test.

In particular, the parser currently only recreates the formatting
elements and queues the insertion tasks before creating the custom
element. This patch ensures that those elements are inserted, so that
we can observe their MutationObserver records in the custom element
constructor.

Bug: 1420810
Change-Id: I35b6a0bf6bfb90144fd132eece0cd12b7491582f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4726970
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1183752}