Replace add eSim button with add eSim entry for QsRevamp
This CL replaced the add eSim button in the mobile header with add eSim
entry at the bottom of the mobile section. This entry row will show up
only under the conditions that the previous eSim button is enabled and
visible. If there's no mobile network set up and the eSim entry is
visible, the "No mobile devices set up" message will not show while the
add eSim entry is showing. This CL also modifies the unit tests to be
compatible with this new entry.

Screenshot:
https://screenshot.googleplex.com/7K6iLLH9fVXTgmk

Bug: b:266761290
Change-Id: I91a0a19db479fa0d48c0638ae81948d12454efed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4711948
Commit-Queue: Sylvie Liu <sylvieliu@chromium.org>
Reviewed-by: Jiaming Cheng <jiamingc@chromium.org>
Reviewed-by: Chad Duffin <chadduffin@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1183765}