Passthrough: Pass UseCmdDecoder flag to renderer for bots
Pass the UseCmdDecoder switch needed for selecting the correct command
decoder to use when running CQ bots based on new feature config.
This is already passed for the gpu_process_host but not for renderer.
This change updates the render_process_host_impl so that the switch can
be used for checking multiplanar shared images feature enabled for
tests that currently override this command line flag.

Bug: 1468989
Change-Id: Ie2d2eae873cb312139090a0631385adbea33c082
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4775054
Reviewed-by: Bo Liu <boliu@chromium.org>
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Saifuddin Hitawala <hitawala@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1183766}