Extensions: WFH: Confirmation dialog for first-time extension file opens
Present a confirmation dialog before opening any of the
extension declared file types for the first time if the extension isn’t
part of the OS (operating system) (or in an allowlist). `Remember my
choice` applies to any declared file type associated with an extension.

In order to "remove extensive business logic from app_service" while
taking advantage of all of the downstream app_service code that helps to
launch apps such as WFH (Web File Handlers), the following flow has been
implemented: app_service:: asks extensions:: if it should ask ui:: for
permissions. If not, app_service:: resumes, otherwise it's halted until
the inquiry is complete. See the attached video demo.

Reuse several parts of the UI where it made sense, such as ensuring that
every file extension is present, instead of just one. If remember is
selected, it is one and done for the entire extension, instead of trying
to manage every single file extension. This CL started out that way, but
the complexity involved in managing every permutation wasn't worth it,
especially given that there is already a prior flow that show each, too.

Demo:
https://drive.google.com/drive/folders/1lHchig6_Ubsytki8cfhvlBNTaXk09e4V

UI:
https://docs.google.com/document/d/1h7ZjryB2zYEjUG9DqPLzAM1iSUXr8ZadUUY02ycExAQ

Design:
https://docs.google.com/document/d/1WTdCYscS5t7bHLyxA1MZ0z8XTmCt8XaIzEBd4sI2lOI

Bug: chromium:1448893
Change-Id: Ia96c0e6fcd69846244ede811ff23622d9d02b5c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4702607
Reviewed-by: Emilia Paz <emiliapaz@chromium.org>
Commit-Queue: Solomon Kinard <solomonkinard@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Reviewed-by: Allen Bauer <kylixrd@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1183796}