[DownloadBubble] Fix dangling pointer when setting checkbox is destroyed
This fixes a dangling pointer to a CheckboxTargeter that was previously
owned by the SuppressBubbleSettingRow. It was being destroyed (as a
non-static class member) before the Checkbox that still had a reference
to it, causing a dangling pointer crash.

This makes it owned by the Checkbox instead.

Bug: 1472833
Change-Id: I1ba00871cf348568824d1973f484ef057948d9eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4781409
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Commit-Queue: Lily Chen <chlily@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1183819}