Add support for Shared Zstd
This CL adds the "zstd-d" token to Accept-Encoding when Compression
Dictionary Transport is enabled and the SharedZstd flag is enabled.
It also adds support for decompressing with Shared Zstd when
Content-Encoding is "zstd-d".

This CL also introduces CreateZstdSourceStreamWithDictionary() which
calls ZSTD_DCtx_loadDictionary_advanced() to use the dictionary.

Bug: 1462543
Change-Id: Ic56caedaf8e76ce008ced631f3d3038cb7d20ee3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4664191
Reviewed-by: Tsuyoshi Horo <horo@chromium.org>
Reviewed-by: Patrick Meenan <pmeenan@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Commit-Queue: Nidhi Jaju <nidhijaju@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1183824}