Change a way the printer is set up when viewing PPD file
This patch changes the method that is used to set a printer up when the
user wants to see the printer's PPD file. In the previous version, the
method ash::printing::SetUpPrinter() is used for setting the printer up.
Unfortunately, this method was designed to be called from PrintPreview
only. This patch solves this problem by calling SetUpPrinter() directly
from CupsPrintersManager.

BUG=b:253246486
TEST=unit_tests

Change-Id: I5a0b48a094f532f03f1247fd20a160c2b50585e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4767996
Reviewed-by: Gavin Williams <gavinwill@chromium.org>
Commit-Queue: Piotr Pawliczek <pawliczek@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1183875}